+++ backend/db/src/main/java/com/bls/patronage/db/dao/FlashcardDAO.java
@@ -1,8 +1,8 @@
 package com.bls.patronage.db.dao;
 
 import com.bls.patronage.db.exception.DataAccessException;
-import com.bls.patronage.db.mappers.FlashcardIdMapper;
-import com.bls.patronage.db.mappers.FlashcardMapper;
+import com.bls.patronage.db.mapper.FlashcardIdMapper;
+import com.bls.patronage.db.mapper.FlashcardMapper;
 import com.bls.patronage.db.model.Flashcard;
 import org.skife.jdbi.v2.sqlobject.Bind;
 import org.skife.jdbi.v2.sqlobject.BindBean;
