+++ backend/db/src/main/java/com/bls/patronage/db/dao/ResultDAO.java
@@ -1,7 +1,7 @@
 package com.bls.patronage.db.dao;
 
 import com.bls.patronage.db.exception.DataAccessException;
-import com.bls.patronage.db.mappers.ResultMapper;
+import com.bls.patronage.db.mapper.ResultMapper;
 import com.bls.patronage.db.model.Result;
 import org.skife.jdbi.v2.sqlobject.Bind;
 import org.skife.jdbi.v2.sqlobject.BindBean;
