+++ backend/db/src/main/java/com/bls/patronage/db/mapper/UserMapper.java
@@ -1,4 +1,4 @@
-package com.bls.patronage.db.mappers;
+package com.bls.patronage.db.mapper;
 
 import com.bls.patronage.db.model.User;
 import org.skife.jdbi.v2.StatementContext;
